add feePayer route and reorganize routes
